refactor: update Rollup config location
