+++ packages/autocomplete-js/rollup.config.js
@@ -1,4 +1,4 @@
-import { createRollupConfigs } from '../../rollup.base.config';
+import { createRollupConfigs } from '../../scripts/rollup/config';
 
 import pkg from './package.json';
